refactor(cli/chat): extract chat::session; chat dm becomes a TUI consumer
Both `chat join` and `chat dm` are now thin wrappers (~140 lines
each) that build a SessionConfig and call session::run. The shared
orchestration (DHT spawn, reader/publisher/nexus/friend/inbox/dht_status
tasks, main select! loop with slash commands, EOF/Ctrl-C drain,
TerminalGuard cleanup) lives in the new chat::session module (~680
lines).

A DM is now structurally a 2-person private chat channel: the only
DM-specific bits are
  * channel_key derived as dm_channel_key(me, them) (vs name+salt)
  * message_key derived as dm_msg_key(ecdh_secret, channel_key)
  * an ephemeral invite-feed-keypair generated at startup
  * an optional initial inbox invite send (--message)
  * a small dm_nudge task that fires send_dm_nudge once per epoch on
    user-typed messages (preserves the original chat dm wire behavior)
  * best-effort empty-payload mutable_put on the invite-feed at
    shutdown (1s timeout) to retract the announce — TTL would expire
    it anyway but a fresh write makes it immediate

Everything else — feed rotation, batching, ordering, dedup, status
bar, scroll region, history replay on resize, Ctrl-C semantics,
inbox monitor, /inbox slash command — flows through unchanged and
DM gets it for free.

## chat dm flag inheritance

`chat dm` now exposes the full chat join flag surface plus its
DM-specific --message:

  --line-mode             auto-detected when stdin or stdout isn't a TTY
  --batch-size N          default 16
  --batch-wait-ms MS      default 50
  --stay-after-eof        keep listening after stdin EOF
  --feed-lifetime MIN     default 60
  --no-nexus / --no-friends / --read-only / --stealth
  --no-inbox              disable the inbox monitor
  --inbox-poll-interval   default 15s
  --message TEXT          DM-only: initial inbox invite lure

Notable behavior changes for `chat dm` users:
  * stdout/stdin TTY → interactive TUI (previously line mode always);
    pipe → line mode automatically (no --line-mode flag required).
  * Outbound messages now batch (mpsc(64), per --batch-size /
    --batch-wait-ms). Interactive typing is bounded by the user; piped
    bursts amortize DHT round trips. Wire format unchanged.
  * Inbox monitor enabled by default — INBOX status bar segment lights
    up on incoming invites; /inbox dumps unread.
  * Bar name uses the canonical NameResolver (friend alias > known
    screen name > vendor name fallback) so it shows
    "DM:alice" / "DM:alice@abc12345" / "DM:tyrannical_elbakyan@abc12345"
    instead of the previous bare-shortkey form.
  * EOF triggers graceful drain instead of indefinite read-only mode
    (matches chat join). Use --stay-after-eof to preserve the old
    behavior.

## Per-post nudge mechanics

The publisher knows nothing about DM. The session loop forwards each
UiInput::Message text into a dedicated mpsc::unbounded channel; a
small dm_nudge task receives, throttles to one nudge per epoch, and
calls send_dm_nudge with the latest text. On shutdown the channel
sender is dropped and the task gets a 1s grace period to finish any
in-flight nudge.

## Net code change

  before: join.rs 666 + dm_cmd.rs 415          = 1,081 lines
  after:  join.rs 141 + dm_cmd.rs 145
                + session.rs 682                = 968 lines

Net -113 lines, but more importantly: every future chat-session
feature lands in one place and benefits both commands automatically.

The deferred 'item 3 — DM as a consumer of the chat TUI framework'
from the original 4-item parent issue list is now done.
