+++ peeroxide-cli/src/cmd/chat/mod.rs
@@ -20,6 +20,7 @@ pub mod probe;
 pub mod profile;
 pub mod publisher;
 pub mod reader;
+pub mod session;
 pub mod tui;
 pub mod wire;
 
@@ -143,7 +144,7 @@ pub async fn run(args: ChatArgs, cfg: &ResolvedConfig) -> i32 {
             .unwrap_or(false);
     match args.command {
         ChatCommands::Join(join_args) => join::run(join_args, cfg, line_mode).await,
-        ChatCommands::Dm(dm_args) => dm_cmd::run(dm_args, cfg).await,
+        ChatCommands::Dm(dm_args) => dm_cmd::run(dm_args, cfg, line_mode).await,
         ChatCommands::Inbox(inbox_args) => inbox_cmd::run(inbox_args, cfg).await,
         ChatCommands::Whoami(args) => run_whoami(args),
         ChatCommands::Profiles { command } => run_profiles(command),
